+++ Gruntfile.js
@@ -1,6 +1,6 @@
 'use strict';
 
-/* eslint no-sync: 0 */
+/* eslint node/no-sync: 0 */
 var path = require('path');
 var fs = require('fs');