+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - **chai** has been upgraded from 4.2.0 to **4.3.4**
 - **grunt** has been upgraded from 1.1.0 to **1.4.1**
 - **grunt-cli** has been upgraded from 1.3.2 to **1.4.3**
+- **grunt-eslint** has been upgraded from 22.0.0 to **23.0.0**
 
 # 7.0.0 / 2020-05-04